\name{LCStr}
\alias{LCStr}
\title{Longest Common SubString}
\usage{
  LCStr(aString, bString)
}
\description{
  pass in 3 parameters, and returns the longest common substring
  arg1: string one
  arg2: string two
  arg3: minimum Length of the common substring.
}
\examples{
  LCStr("AABCC","BACCABC",1) returns "ABC"

}
